backend/requirements.txt
asgiref==3.9.1
attrs==25.3.0
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.0
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.5
django-ratelimit==4.1.0
django-redis==6.0.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.26.5
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
pika==1.3.2
pillow==11.3.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
rpds-py==0.26.0
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
